+++ creep.action.building.js
@@ -1,6 +1,5 @@
-var action = new MODULES.creep.Action();
+var action = new Creep.Action('building');
 
-action.name = 'building';
 action.ignoreCreeps = true;
 
 action.isValidAction = function(creep){
