+++ creep.action.feeding.js
@@ -1,6 +1,6 @@
-var action = new MODULES.creep.Action();
+var action = new Creep.Action('feeding');
 
-action.name = 'feeding';
+action.maxPerTarget = 1;
 
 action.isValidAction = function(creep){
     return ( creep.carry.energy > 0 && creep.room.energyAvailable < creep.room.energyCapacityAvailable );
@@ -11,7 +11,7 @@ action.isValidTarget = function(target){
 action.isAddableAction = function(creep){
     return (!creep.room.activities[this.name] || creep.room.activities[this.name] < (creep.room.maxPerJob * (creep.room.relativeEnergyAvailable < HIVE_ENERGY_URGENT ? 2 : 1)) );
 };
-action.maxPerTarget = 1;
+
 action.newTarget = function(creep){
     var self = this;
     return creep.pos.findClosestByPath(FIND_STRUCTURES, {
