+++ creep.action.withdrawing.js
@@ -1,6 +1,6 @@
-var action = new MODULES.creep.Action();
+var action = new Creep.Action('withdrawing');
 
-action.name = 'withdrawing';
+action.maxPerTarget = 2;
 
 action.isValidAction = function(creep){
     return ( creep.room.storage && _.sum(creep.carry) < creep.carryCapacity && 
@@ -10,7 +10,6 @@ action.isValidAction = function(creep){
 action.isValidTarget = function(target){
     return ( (target != null) && (target.store != null) && (target.store.energy > 0) );
 };  
-action.maxPerTarget = 2;
 action.newTarget = function(creep){
     return creep.room.storage;
 };
