+++ creep.behaviour.claimer.js
@@ -1,4 +1,4 @@
-var behaviour = new MODULES.creep.Behaviour();
+var behaviour = new Creep.Behaviour('claimer');
 
 behaviour.run = function(creep) {
     var flag;
