+++ creep.setup.pioneer.js
@@ -1,6 +1,5 @@
-var setup = new MODULES.creep.Setup();
+var setup = new Creep.Setup('pioneer');
 
-setup.type = 'pioneer';
 setup.multiBody = [CARRY, WORK, MOVE];
 setup.fixedBody = [CARRY, WORK, MOVE, CARRY, WORK, MOVE];
 setup.minAbsEnergyAvailable = 400;
