+++ creep.setup.privateer.js
@@ -1,6 +1,5 @@
-var setup = new MODULES.creep.Setup();
+var setup = new Creep.Setup('privateer');
 
-setup.type = 'privateer';
 setup.multiBody = [CARRY, WORK, MOVE];
 setup.fixedBody = [CARRY, WORK, MOVE, CARRY, WORK, MOVE];
 setup.minAbsEnergyAvailable = 400;
@@ -12,7 +11,7 @@ setup.minEnergyAvailable = function(){
     return 0.8;
 };
 setup.maxCount = function(spawn){
-    return _.filter(Game.flags, FLAG_COLOR.invade.exploit.filter).length * 5;
+    return _.filter(Game.flags, FLAG_COLOR.invade.exploit.filter).length * 6;
 }
 setup.maxWeight = function(spawn){
     return null;
