+++ global.js
@@ -64,51 +64,6 @@ var mod = {
                 }
             }
         };
-        global.MODULES = {};
-        global.MODULES.creep = require('creep');
-        global.MODULES.creep.action = {
-            building: require('creep.action.building'), 
-            claiming: require('creep.action.claiming'), 
-            defending: require('creep.action.defending'),
-            feeding: require('creep.action.feeding'), 
-            fueling: require('creep.action.fueling'), 
-            guarding: require('creep.action.guarding'), 
-            harvesting: require('creep.action.harvesting'),
-            healing: require('creep.action.healing'),
-            idle: require('creep.action.idle'),
-            invading: require('creep.action.invading'),
-            picking: require('creep.action.picking'), 
-            repairing: require('creep.action.repairing'), 
-            settling: require('creep.action.settling'), 
-            storing: require('creep.action.storing'), 
-            upgrading: require('creep.action.upgrading'), 
-            withdrawing: require('creep.action.withdrawing')
-        };
-        global.MODULES.creep.behaviour = {
-            claimer: require('creep.behaviour.claimer'),
-            healer: require('creep.behaviour.healer'),
-            melee: require('creep.behaviour.melee'),
-            pioneer: require('creep.behaviour.pioneer'),
-            privateer: require('creep.behaviour.privateer'),
-            ranger: require('creep.behaviour.ranger'),
-            worker: require('creep.behaviour.worker')
-        };
-        global.MODULES.creep.setup = {
-            claimer: require('creep.setup.claimer'),
-            healer: require('creep.setup.healer'), 
-            melee: require('creep.setup.melee'),
-            pioneer: require('creep.setup.pioneer'),
-            privateer: require('creep.setup.privateer'),
-            ranger: require('creep.setup.ranger'),
-            worker: require('creep.setup.worker')
-        };
-        
-        global.MODULES.extensions = require('extensions');
-        global.MODULES.population = require('population');
-        global.MODULES.room = require('room');
-        global.MODULES.spawn = require('spawn'); 
-        global.MODULES.tower = require('tower');
-
         global.PART_COSTS = {
             work: 100,
             carry: 50,
@@ -144,6 +99,13 @@ var mod = {
             console.log( message );
             Game.notify( message, 120 );
         };
+        global.MODULES = {};
+        global.MODULES.creep = require('creep');        
+        global.MODULES.extensions = require('extensions');
+        global.MODULES.population = require('population');
+        global.MODULES.room = require('room');
+        global.MODULES.spawn = require('spawn'); 
+        global.MODULES.tower = require('tower');
     }
 }
