+++ spawn.js
@@ -18,7 +18,7 @@ var mod = {
         ].forEach(function(set) {
             if( !spawn.busy && set.isValidSetup(spawn) ){
                 var params =  set.buildParams(spawn);
-                //if( DEBUG ) console.log(spawn.name + ' > Spawning new ' + params.setup + ' @ ' + (Game.population[params.setup] ? Game.population[params.setup].count : 0) + ' total already existing');
+                
                 var newName = spawn.createCreep(params.parts, params.id, params);
                 if( params.id == newName ){
                     spawn.busy = true;
@@ -28,7 +28,6 @@ var mod = {
                 console.log(params.id == newName || ERROR_CODE(newName) === undefined ? 
                     spawn.name + ' > Good morning ' + newName + '!': 
                     spawn.name + ' > Offspring failed: ' + ERROR_CODE(newName));
-                //if( DEBUG ) console.log(spawn.name + ' > New total population of ' + params.setup + ': ' + (Game.population[params.setup] ? Game.population[params.setup].count : 0) + '<br/> ');
             }
         });
     }
